Add HAProxy
Adds a connection queue on (default) port 10001 for TTS-operations,
and keeps the raw port 10000 open for all the other operations.
Also runs a HAProxy statistics page on (default) port :10002/stats
for monitoring the current queue load.

HAProxy is configurable via environment variables that are documented
in the /blubber-readme.md file.

Change-Id: Ib6cf03eca7b6d5374378d5be8eae6f49d5008298